# Performance Optimization - Phase 1: Array Pre-allocation and Vectorization