## Summary
- Pre-allocate arrays in `_forward` function to eliminate expensive `list.append()` + `np.vstack()` pattern
- Vectorize selection matrix creation using boolean indexing instead of loops
- Add matrix caching for frequently used w and F matrices  
- Add comprehensive benchmark script for performance testing

## Performance Improvements
- **6-7% speedup** on 50x5 dataset (LocalLevel: 0.052s → 0.049s, LocalTrend: 0.090s → 0.084s)
- Eliminates memory allocation overhead in tight loops
- Reduces function call overhead with cached matrices

## Test Results
- ✅ All 37 tests pass
- ✅ Identical behavior maintained
- ✅ Code quality checks pass

## Next Steps
Phase 2 will add Numba JIT compilation for larger performance gains.

🤖 Generated with [Claude Code](https://claude.ai/code)